+++ examples/storybook/src/shared/utils/linkPreviewUtil.ts
@@ -17,7 +17,7 @@ export const mockFetchUrlPreviewData = () => {
   };
 };
 
-const embedTypes = ['instagram', 'twitter', 'youtube', 'tiktok'];
+const embedTypes = ['instagram', 'twitter', 'youtube', 'tiktok', 'spotify'];
 
 const mockLinkPreviewResults = [
   {
@@ -84,6 +84,10 @@ const mockOembedResults = {
     thumbnail_url: 'https://www.tiktok.com/@dayana_yaish',
     provider_url: 'https://www.tiktok.com',
     html: '<blockquote class="tiktok-embed" cite="https://www.tiktok.com/@dayana_yaish/video/6794733203912445189" data-video-id="6794733203912445189" style="max-width: 605px;min-width: 325px;" > <section> <a target="_blank" title="@dayana_yaish" href="https://www.tiktok.com/@dayana_yaish">@dayana_yaish</a> <p><a title="ohnanachallenge" target="_blank" href="https://www.tiktok.com/tag/ohnanachallenge">#ohnanachallenge</a> <a title="ohnachallenge" target="_blank" href="https://www.tiktok.com/tag/ohnachallenge">#ohnachallenge</a> <a title="ohnanana" target="_blank" href="https://www.tiktok.com/tag/ohnanana">#ohnanana</a> <a title="lablife" target="_blank" href="https://www.tiktok.com/tag/lablife">#lablife</a> @likgamaev</p> <a target="_blank" title="♬ Oh Nanana - Remix - dj 6rb & bonde r300" href="https://www.tiktok.com/music/Oh-Nanana-Remix-6624126913629653766">♬ Oh Nanana - Remix - dj 6rb & bonde r300</a> </section> </blockquote> <script async src="https://www.tiktok.com/embed.js"></script>',
+  },
+  spotify: {
+    provider_url: 'https://www.spotify.com',
+    html: '<iframe width="100%" height="80" title="Spotify Embed: Grateful"  frameborder="0" allowfullscreen allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture" src="https://open.spotify.com/embed/track/2wmZFiWL7MKt2vQe6DPanz?si=aeab9fbc67954570&utm_source=oembed"></iframe>',
   },
 };