+++ packages/common/web/src/test/contentStateValidator.spec.ts
@@ -2,6 +2,7 @@
 import { isValidEditorData } from '../../lib/contentStateValidator';
 import EmojiContent from './fixtures/emojiContent.json';
 import PollsContent from './fixtures/pollsContent.json';
+import audioContent from './fixtures/audioContent.json';
 describe('Content State Validator', () => {
   it('should return invalid result', async () => {
     const validation = isValidEditorData({ what: 'hey!' });
@@ -168,6 +169,10 @@ describe('Content State Validator', () => {
 
   it('should return valid for poll content', async () => {
     const validation = isValidEditorData(PollsContent);
+    expect(validation.valid).toEqual(true);
+  });
+  it('should return valid for audio content', async () => {
+    const validation = isValidEditorData(audioContent);
     expect(validation.valid).toEqual(true);
   });
 });