+++ packages/plugin-audio/web/src/components/index.ts
@@ -1,5 +1,4 @@
 export { default as AudioActionIcons } from './AudioActionIcons';
-// export { default as AudioSlider } from './AudioSlider';
 export { default as AudioDetails } from './AudioDetails';
 export { default as AudioDuration } from './AudioDuration';
 export { default as AudioImage } from './AudioImage';